requirements.txt
altair>=4.2.0
attrs>=22.2.0
beautifulsoup4>=4.12.0
blinker>=1.6.0
cachetools>=5.3.0
camelcase>=0.2
certifi>=2023.5.0
charset-normalizer>=3.1.0
click>=8.1.3
colorama>=0.4.6
contourpy>=1.0.7
cycler>=0.11.0
decorator>=5.1.1
distlib>=0.3.6
dnspython>=2.3.0
eli5>=0.13.0
entrypoints>=0.4
Faker>=18.9.0
favicon>=0.7.0
filelock>=3.10.0
fonttools>=4.39.0
gitdb>=4.0.10
GitPython>=3.1.31
google-api-core>=2.11.0
google-api-python-client>=2.86.0
google-auth>=2.18.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
googleapis-common-protos>=1.59.0
graphviz>=0.20.1
greenlet>=2.0.2
htbuilder>=0.6.1
httplib2>=0.22.0
idna>=3.4
importlib-metadata>=6.6.0
Jinja2>=3.1.2
joblib>=1.2.0
jsonschema>=4.17.3
kiwisolver>=1.4.4
lxml>=4.9.2
Markdown>=3.4.3
markdown-it-py>=2.2.0
markdownlit>=0.0.7
MarkupSafe>=2.1.2
matplotlib>=3.7.1
mdurl>=0.1.2
more-itertools>=9.1.0
mysql-connector-python>=8.0.32
numpy>=1.24.2
oauthlib>=3.2.2
packaging>=23.0
pandas>=1.5.3
Pillow>=9.4.0
platformdirs>=3.1.1
plotly>=5.14.1
protobuf>=3.20.3
psycopg2-binary>=2.9.6
pyarrow>=12.0.0
pyasn1>=0.5.0
pyasn1-modules>=0.3.0
pydeck>=0.8.0
Pygments>=2.15.1
pymdown-extensions>=10.0.1
pymongo>=4.3.3
Pympler>=1.0.1
pyparsing>=3.0.9
pyrsistent>=0.19.3
python-dateutil>=2.8.2
pytz>=2022.7.1
pytz-deprecation-shim>=0.1.0
PyYAML>=6.0
requests>=2.30.0
requests-oauthlib>=1.3.1
rich>=13.3.5
rsa>=4.9
scikit-learn>=1.2.2
scipy>=1.10.1
seaborn>=0.12.2
six>=1.16.0
smmap>=5.0.0
soupsieve>=2.4.1
SQLAlchemy>=2.0.15
st-annotated-text>=4.0.0
statsmodels>=0.14.0
streamlit>=1.22.0
streamlit-camera-input-live>=0.2.0
streamlit-card>=0.0.4
streamlit-embedcode>=0.1.2
streamlit-extras>=0.2.7
streamlit-faker>=0.0.2
streamlit-image-coordinates>=0.1.3
streamlit-keyup>=0.2.0
streamlit-toggle-switch>=1.0.2
streamlit-vertical-slider>=1.0.2
tabulate>=0.9.0
tenacity>=8.2.2
threadpoolctl>=3.1.0
toml>=0.10.2
toolz>=0.12.0
tornado>=6.3.1
typing_extensions>=4.5.0
tzdata>=2023.3
tzlocal>=4.3
uritemplate>=4.1.1
urllib3>=1.26.15
validators>=0.20.0
virtualenv>=20.21.0
watchdog>=3.0.0
xlrd3>=1.1.0
zipp>=3.15.0
python-dotenv>=1.0.0
